[wip]mb/dell: add Dell Optiplex XE2
Adds support for the Dell Optiplex XE2 utilizing the SCH555x ec sources. The code is similar enough to the Dell Optiplex 9020 that I think this should actually be merged with the sources for the Optiplex 9020, but I just wanted to put this up here so it's publicly accessible. Compiled and tested with SeaBIOS and Tianocore booting into Arch Linux kernel 5.13.12. The Haswell mrc.bin blob is required.

Tested working
- SATA SSD
- PCI-e graphics card via PRIME
- libgfxinit
- External flashing
- Audio jack
- USB
- VGA

Untested
- SATA DVD drive
- PCI-e x1
- Second PCI-e x16
- DisplayPort
- Serial port
- PS/2 keyboard and mouse
- Gigabit Ethernet

Signed-off-by: Justin Wu <amersel@runbox.com>
Change-Id: Iab9f8a4611bb76201463839d0251833d7880106b