acpi: Generate power resource name instead of default "PRIC"
This change uses a static 8-bit counter in
`acpi_device_add_power_res()` to generate the name of power resource
object rather using a default name "PRIC". This makes it easier to
identify which power resource Linux kernel logs are referring to. If
more than 256 power resources are used in the system, then the counter
will wrap around to 0. However, 256 seems to be a large enough number
for the power resource count.

TEST=Verified that Power Resources are named as expected:
```
dmesg | grep ACPI | grep PR
[    0.550921] ACPI: Power Resource [PR00] (on)
[    0.869960] ACPI: Power Resource [PR01] (on)
[    1.013973] ACPI: Power Resource [PR02] (on)
```

No new ACPI errors are seen in dmesg on brya.

Change-Id: Ia18f7177b03821ce0f8c989ae5d258f2f83517a5
Signed-off-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/57650
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>